Adding my basic source code.
Hey Michael,
This is the code I have so far and it essentially runs as intended. I still need to edit it so it prints an error message if there is no solution and make stylistic edits but I wanted you to have something to work with. Let me know if you have any questions or suggestions.